refactor code regarding updating posts
