+++ post/index.html
@@ -268,7 +268,9 @@
         <!-- Edit post  -->
         <section class="px-3 w-full">
           <div class="flex flex-col gap-4 mt-4 max-w-xl mx-auto">
-            <label for="post-title" class="text-center text-3xl font-semibold">Edit post</label>
+            <label id="edit-label" for="post-title" class="text-center text-3xl font-semibold"
+              >Edit post</label
+            >
             <div id="message-container"></div>
             <!-- Form -->
             <form id="edit-post-form" enctype="multipart/form-data" class="flex flex-col gap-4 h-full">
@@ -279,6 +281,7 @@
                     type="text"
                     name="title"
                     id="post-title"
+                    placeholder="Title"
                     required
                     minlength="3"
                     class="appearance-none outline-none pt-3 px-3 bg-transparent w-full h-10 placeholder:text-gray-600 dark:placeholder:text-gray-400" />
@@ -289,7 +292,7 @@
                   <textarea
                     name="body"
                     id="post-body"
-                    placeholder="body"
+                    placeholder="Body text"
                     class="appearance-none outline-none pt-3 px-3 bg-transparent w-full placeholder:text-gray-600 dark:placeholder:text-gray-400 h-96"></textarea>
                   <!-- Image Upload Button -->
                   <div class="flex items-center">
