+++ src/css/styles.css
@@ -926,14 +926,6 @@ video {
   margin-top: auto;
 }
 
-.mt-24 {
-  margin-top: 6rem;
-}
-
-.mt-96 {
-  margin-top: 24rem;
-}
-
 .inline-block {
   display: inline-block;
 }
@@ -1003,6 +995,10 @@ video {
   height: 2rem;
 }
 
+.h-96 {
+  height: 24rem;
+}
+
 .h-fit {
   height: -moz-fit-content;
   height: fit-content;
@@ -1014,30 +1010,6 @@ video {
 
 .h-screen {
   height: 100vh;
-}
-
-.h-1 {
-  height: 0.25rem;
-}
-
-.h-64 {
-  height: 16rem;
-}
-
-.h-44 {
-  height: 11rem;
-}
-
-.h-72 {
-  height: 18rem;
-}
-
-.h-96 {
-  height: 24rem;
-}
-
-.max-h-screen {
-  max-height: 100vh;
 }
 
 .min-h-full {
@@ -1081,10 +1053,6 @@ video {
   width: 100vw;
 }
 
-.min-w-full {
-  min-width: 100%;
-}
-
 .max-w-2xl {
   max-width: 42rem;
 }
@@ -1095,10 +1063,6 @@ video {
 
 .max-w-xl {
   max-width: 36rem;
-}
-
-.transform {
-  transform: translate(var(--tw-translate-x), var(--tw-translate-y)) rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y)) scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
 }
 
 @keyframes spin {
@@ -1183,10 +1147,6 @@ video {
 
 .gap-8 {
   gap: 2rem;
-}
-
-.self-center {
-  align-self: center;
 }
 
 .scroll-smooth {
@@ -24411,10 +24371,6 @@ video {
      object-fit: cover;
 }
 
-.p-6 {
-  padding: 1.5rem;
-}
-
 .px-2 {
   padding-left: 0.5rem;
   padding-right: 0.5rem;
@@ -24448,11 +24404,6 @@ video {
 .py-2 {
   padding-top: 0.5rem;
   padding-bottom: 0.5rem;
-}
-
-.py-6 {
-  padding-top: 1.5rem;
-  padding-bottom: 1.5rem;
 }
 
 .pb-1 {
@@ -47634,12 +47585,6 @@ video {
   box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
 }
 
-.shadow-lg {
-  --tw-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);
-  --tw-shadow-colored: 0 10px 15px -3px var(--tw-shadow-color), 0 4px 6px -4px var(--tw-shadow-color);
-  box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
-}
-
 .outline-none {
   outline: 2px solid transparent;
   outline-offset: 2px;
@@ -47850,10 +47795,6 @@ input:-webkit-autofill,
     height: 11rem;
   }
 
-  .sm\:h-96 {
-    height: 24rem;
-  }
-
   .sm\:w-10 {
     width: 2.5rem;
   }
@@ -47903,34 +47844,6 @@ input:-webkit-autofill,
     height: 1.75rem;
   }
 
-  .lg\:h-64 {
-    height: 16rem;
-  }
-
-  .lg\:h-screen {
-    height: 100vh;
-  }
-
-  .lg\:h-14 {
-    height: 3.5rem;
-  }
-
-  .lg\:h-12 {
-    height: 3rem;
-  }
-
-  .lg\:h-3\/4 {
-    height: 75%;
-  }
-
-  .lg\:h-1\/2 {
-    height: 50%;
-  }
-
-  .lg\:h-96 {
-    height: 24rem;
-  }
-
   .lg\:w-fit {
     width: -moz-fit-content;
     width: fit-content;
@@ -47938,14 +47851,6 @@ input:-webkit-autofill,
 
   .lg\:focus-within\:h-20:focus-within {
     height: 5rem;
-  }
-
-  .lg\:focus-within\:h-60:focus-within {
-    height: 15rem;
-  }
-
-  .lg\:focus-within\:h-64:focus-within {
-    height: 16rem;
   }
 
   .lg\:hover\:stroke-2:hover {
